ProyectoCultura/requirements.txt
Django==1.7.11
django-braces==1.10.0
djrill==2.1.0
mysqlclient==1.3.9
Pillow==3.4.2
requests==2.11.1
Unipath==1.1
